requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
bcrypt==4.2.0
blinker==1.8.2
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
dnspython==2.7.0
dotenv==0.9.9
faiss-cpu==1.11.0
filelock==3.18.0
filetype==1.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-PyMongo==2.3.0
flask-redis==0.4.0
Flask-SQLAlchemy==3.1.1
flet==0.27.6
flet-desktop==0.27.6
fonttools==4.56.0
frozenlist==1.6.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.58
langchain-google-genai==2.1.4
langchain-text-splitters==0.3.8
langgraph==0.4.2
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.42
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
mysql-connector-python==9.1.0
networkx==3.4.2
numpy==2.2.0
oauthlib==3.2.2
ollama==0.4.5
opencv-python==4.11.0.86
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.1
paho-mqtt==2.1.0
paramiko==3.5.0
pillow==11.1.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.27.2
PyJWT==2.9.0
pymongo==4.10.1
PyMuPDF==1.25.5
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==5.2.0
repath==0.9.0
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.6.0
scipy==1.14.1
scp==0.15.0
setuptools==80.3.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.5.0
torch==2.7.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.0.6
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0




#rag
# RAG Intelligent Teaching System - Dependencies
# Core packages for PDF processing, vector database, and AI integration

# Core frameworks
numpy>=1.21.0
pandas>=1.3.0

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.7.0
unstructured>=0.10.0
unstructured[pdf]>=0.10.0

# Text processing and NLP
nltk>=3.8
jieba>=0.42.1
sentence-transformers>=2.2.0
transformers>=4.21.0

# Vector databases
chromadb>=0.4.0
faiss-cpu>=1.7.4

# HTTP clients and APIs
requests>=2.28.0
httpx>=0.24.0

# Machine learning and deep learning
torch>=1.12.0
torchvision>=0.13.0
scikit-learn>=1.1.0

# Data processing utilities
tqdm>=4.64.0
python-dotenv>=0.19.0

# Logging and configuration
coloredlogs>=15.0
pyyaml>=6.0

# Development and testing tools
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

